feat: sigagg (#426)
* Dutydb first attempt

* style improvements

* await_data method

* rwlock instead of mutex

* renamed variables

* State struct methods

* improved test

* Locking state after adding duty to the deadliner in store function

* fixed Dead clash check in store_agg_attestation

* store_agg_attestation inserting only when not slot present

* fmt

* return early on DeprecatedDutyBuilderProposer

* redundant enable

* removed Option from the State, deadliner field.

* comment ported from charon

* additional, missing tests

* rephrased error messages

* Fixed bug with second aggregations on the same slot

* clippy

* fixed #7

* better implementation of the #1

* Impl Drop for MemDB

* fixed comment 2

* More detailed errors

* fix comment 11

* fix unneeded pub(crate)

* comment

* additional tests

* removed unused, renamed variables

* comment fixed

* more comments fixed

* comment updated

* more uses, removed export o Result

* Checkpoint::default()

* renamed variables

* sigagg first attempt

* More error cases, additional tests

* Safe conversion from u64 to u8

* Better erros for aggregate_one

* comment for aggregate_one

* info_span as in charon

* unneeded Sync in functions results

* set passed as ref to aggregate function

* pub result

* fixed top comment

* Additional tests suggested by claude

* additional comment

* Signature from the crypto crate

* review

* share_idx in ParSignedData changed to u8

* fmt

* clippy

* adjusted spans to match charon's logging

* Unneeded dependency in core/Cargo.toml

* peer index and share index as u64 values

* imported tracing::error

* fixed cargo machete

+++ crates/cli/src/commands/test/peers.rs
@@ -413,7 +413,7 @@ fn parse_peers(enr_strings: &[String]) -> Result<Vec<Peer>> {
         .enumerate()
         .map(|(i, s)| {
             let record = Record::try_from(s.as_str())?;
-            Ok(Peer::from_enr(&record, i)?)
+            Ok(Peer::from_enr(&record, i as u64)?)
         })
         .collect()
 }
